refactor: Van der Waals and broadcasting tests
